update the js
Need to fix the collision part of the rocket and the comet for a win or lose situation. The problem is js if doesn't let us work with vh or vw units thus making that logic quite tough for me as I am quite beginner in js at this moment.